@@ -5,14 +5,14 @@ homepage: https://github.com/lukasgit/flutter_contacts
 maintainer: Lukas Dickie (@lukasgit)
 
 environment:
-  sdk: '>=2.12.0-0 <3.0.0'
-  flutter: '>=1.12.0'
+  sdk: ^3.5.0
+  flutter: ^3.24.3
 
 dependencies:
   flutter:
     sdk: flutter
-  collection: ^1.15.0
-  quiver: ^3.0.0
+  collection: 1.18.0
+  quiver: ^3.2.2
 
 flutter:
   plugin:
@@ -26,4 +26,4 @@ flutter:
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  test: ^1.3.0
+  test: 1.25.7
